ai_layer/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
anthropic
python-dotenv
pypdf